notification feature without real time update

+++ src/app/api/posts/[postId]/likes/route.ts
@@ -60,7 +60,19 @@ export async function POST(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.upsert({
+    // the user has liked certain post.
+
+    const post = await prisma.post.findUnique({
+      where: {
+        id: postId,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "User not found" }, { status: 404 });
+    }
+
+    const upsertLike = prisma.like.upsert({
       where: {
         postId_userId: {
           postId: postId,
@@ -73,6 +85,20 @@ export async function POST(
       },
       update: {},
     });
+
+    const createNotification = prisma.notification.create({
+      data: {
+        issuerId: loggedInUser.id,
+        recipientId: post.userId,
+        postId: post.id,
+        type: "LIKE",
+      },
+    });
+
+    await prisma.$transaction([
+      upsertLike,
+      ...(post.userId !== loggedInUser.id ? [createNotification] : []),
+    ]);
 
     return new Response();
   } catch (error) {
@@ -92,12 +118,37 @@ export async function DELETE(
       return Response.json({ error: "Unauthorized" }, { status: 401 });
     }
 
-    await prisma.like.deleteMany({
+    const post = await prisma.post.findUnique({
+      where: {
+        id: postId,
+      },
+    });
+
+    if (!post) {
+      return Response.json({ error: "User not found" }, { status: 404 });
+    }
+
+    const deleteLike = prisma.like.deleteMany({
       where: {
         postId: postId,
         userId: loggedInUser.id,
       },
     });
+
+    // we dont have the id of the like to be deleted. Delete wont work without it. therefore using deleteMany
+    const deleteNotification = prisma.notification.deleteMany({
+      where: {
+        issuerId: loggedInUser.id,
+        recipientId: post.userId,
+        type: "LIKE",
+        postId: post.id,
+      },
+    });
+
+    // deleteMany need not be in a condition because if there is no notification if will just be ignored.
+    // notification is not created when the user likes his own post, therefore there is nothing to delete. hence no condition needed.
+
+    await prisma.$transaction([deleteLike, deleteNotification]);
 
     return new Response();
   } catch (error) {
